requirements.txt
openai>=1.40.0
pyyaml>=6.0
tqdm>=4.66.0
